+++ server/api/users.js
@@ -1,8 +1,19 @@
 const router = require('express').Router()
 const {User} = require('../db/models')
+const {Storage} = require('@google-cloud/storage');
 const multer = require('multer')
 module.exports = router
 
+const gcs = new Storage({
+  projectId: process.env.GOOGLE_CLOUD_PROJECT,
+  keyFilename: './KEY.JSON'
+});
+const bucketName = process.env.GCLOUD_STORAGE_BUCKET;
+const bucket = gcs.bucket(bucketName);
+
+function getPublicUrl(filename) {
+  return 'https://storage.googleapis.com/' + bucketName + '/' + filename;
+}
 
 function checkFileType(file, cb){
   // Allowed ext
@@ -19,24 +30,31 @@ function checkFileType(file, cb){
   }
 };
 
-const storage = multer.diskStorage({
-  destination: function(req, file, cb) {
-    cb(null, './user/profileUpload/');
-  },
-  filename: function(req, file, cb) {
-    cb(null, req.params.userName + '-' + file.originalname);
-  }
-});
-
 const upload = multer({
-  storage: storage,
-  // limits: {
-  //   fileSize: 1024 * 1024 * 5
-  // },
+  storage: multer.MemoryStorage,
   fileFilter: function(req, file, cb){
     checkFileType(file, cb);
   }
-});
+})
+
+// const storage = multer.diskStorage({
+//   destination: function(req, file, cb) {
+//     cb(null, './user/profileUpload/');
+//   },
+//   filename: function(req, file, cb) {
+//     cb(null, req.params.userName + '-' + file.originalname);
+//   }
+// });
+
+// const upload = multer({
+//   storage: storage,
+//   // limits: {
+//   //   fileSize: 1024 * 1024 * 5
+//   // },
+//   fileFilter: function(req, file, cb){
+//     checkFileType(file, cb);
+//   }
+// });
 
 router.get('/', async (req, res, next) => {
   try {
@@ -80,6 +98,31 @@ router.put('/:id', async (req, res, next) => {
 })
 
 router.post("/:id", upload.single('profilePic'), (req, res, next) => {
+  if(!req.file) return next();
+  const gcsname = req.params.userName + '-' + req.file.originalname;
+  const blob = bucket.file(gcsname);
+  const stream = blob.createWriteStream({
+    metadata: {
+      contentType: req.file.mimetype
+    }
+  });
+
+  stream.on('error', (err) => {
+    req.file.cloudStorageError = err;
+    next(err);
+  });
+
+  stream.on('finish', () => {
+    req.file.cloudStorageObject = gcsname;
+    req.file.cloudStoragePublicUrl = getPublicUrl(gcsname);
+    next();
+  });
+
+  stream.end(req.file.buffer);
+  const path = 'https://storage.googleapis.com/'+ bucketName+ '/' + gcsname;
+  blob.makePublic().then(() => {
+    res.status(200).send('Success!\n profilePic uploaded to:' + path);
+  });
   User.create({userName: req.body.userName, email: req.body.email, profilePicture:req.body.path})
     .then(result => {
       console.log(result);
